RcppDate-package {RcppDate}	R Documentation

'date' C++ Header Library for Date and Time Functionality

Description

A header-only C++ library is provided with support for dates, time zones, ISO weeks, Julian dates, and Islamic dates. 'date' offers extensive date and time functionality for the C++11, C++14 and C++17 standards and was written by Howard Hinnant and released under the MIT license. A slightly modified version has been accepted (along with 'tz.h') as part of C++20. This package regroups all header files from the upstream repository by Howard Hinnant so that other R packages can use them in their C++ code. At present, few of the types have explicit 'Rcpp' wrappers though these may be added as needed.

Package Content

Index of help topics:

RcppDate-package        'date' C++ Header Library for Date and Time
                        Functionality

Maintainer

Dirk Eddelbuettel <edd@debian.org>

Author(s)

Dirk Eddelbuettel [aut, cre] (ORCID: <https://orcid.org/0000-0001-6419-907X>), Howard Hinnant [aut] (author of 'date' library)

Examples

## see the source files in the examples/ directory of the package
## check for (optional, only in Suggests:) Rcpp, and ensure we are
## not on Windows as this runs up to the ten second time limit
if (requireNamespace("Rcpp", quietly=TRUE) && (.Platform$OS.type != "windows")) {
    Rcpp::sourceCpp(system.file("examples", "year_month_day.cpp", package="RcppDate"))
}

[Package RcppDate version 0.0.6 Index]